requirements.txt
flask
python-dotenv
groq
requests
dnspython
beautifulsoup4
reportlab
python-whois